feat(ScreenSize.jsx, utils.js, README.md): add support for detecting OS name
Add support for detecting the OS name in the `ScreenSize` component by adding a new function `getOSName` to the `utils.js` file. The function detects the OS name based on the user agent and platform. The detected OS name is then added to the `screen` object in the `ScreenSize` component.

Update the `ScreenSize` component to call the `getOSName` function and add the detected OS name to the `screen` object.

Update the `README.md` file to document the new `os` property in the `screen` object.